requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.2
django-celery-beat==2.5.0
django-crum==0.7.9
django-environ==0.10.0
django-extensions==3.2.3
django-redis==5.4.0
django-rest-framework==0.1.0
django-timezone-field==5.1
djangorestframework==3.14.0
Faker==18.10.1
gunicorn==20.1.0
kombu==5.3.1
prompt-toolkit==3.0.38
psycopg2==2.9.6
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.3.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
